ovn: move gress policy specific code to gress_policy.go
Straight copy & paste with no other changes.
